+++ pyspider/database/sqlalchemy/resultdb.py
@@ -23,7 +23,7 @@ class ResultDB(SplitTableMixin, BaseResultDB):
                            Column('taskid', String(64), primary_key=True, nullable=False),
                            Column('url', String(1024)),
                            Column('result', LargeBinary),
-                           Column('updatetime', Float(16, 4))
+                           Column('updatetime', Float(16))
                            )
         self.engine = create_engine(url)
